Only include Traceur runtime if used
This changes, based on the new version of Traceur, to only include the
runtime if the source is using the runtime. Because the runtime no
longer uses the `assertObject` calls it used to, for simple scenarios
such as only importing/exporting the runtime is not needed. This changes
`has_es6` to instead of testing the format, test for the presence of
`$traceurRuntime` in the source. Fixes #75 and #99